do not reuse defaultHttpClient in tilgangersøknad
since ktor plugin installs are appending, defaultHttpClient causes application/json
header to be added.
Altinn API, although documented as hal+json will prioritize application/json if the header is present.
The response payload for json is different, and undocumented for application/json.
Rewriting the client now, when the api will soon be removed by altinn, is not worth it..